frontend/requirements.txt
streamlit
requests
pandas
numpy
matplotlib
seaborn
Pillow